last_position percentage type error int -> float

+++ wallet.py
@@ -409,9 +409,9 @@ class Wallet:
                 h += f"Pnl: {p['info']['pnl']} netProfit: {p['info']['netProfit']}\n"
                 h += f"Openfee: {p['info']['openFee']} Closefee: {p['info']['closeFee']}\n Funding fee: {p['info']['totalFunding']}"
 
-            open_price = int(p['info']['openAvgPrice'])
-            close_price = int(p['info']['closeAvgPrice'])
-            percentage_pnl = (close_price - open_price) / open_price - (close_price * int(p['info']['closeFee']) + open_price * int(p['info']['openFee']))
+            open_price = float(p['info']['openAvgPrice'])
+            close_price = float(p['info']['closeAvgPrice'])
+            percentage_pnl = (close_price - open_price) / open_price - (close_price * float(p['info']['closeFee']) + open_price * float(p['info']['openFee']))
             return h, percentage_pnl
 
         except Exception as e:
